Crate dfn_protobuf

Structs

ProtoBuf

Traits

ToProto
This is deliberately less flexible than From/TryFrom because they can have multiple types they can be transformed into, preventing type inference. Each type can only have one proto type it maps to.

Functions

protobuf
This is the witness for protobuf types (types with a prost::Message implementation) and types that convert to a protobuf type (types with a ToProto implementation).